requirements.txt
pygame-ce
pygame_widgets